♻️ refactor(core,runtime,plugin,sdk,tools,docs): 实现 PolicyEngine 三态契约与 ApprovalBroker，修复代码审查问题
📋 变更概述
实现 Phase 3 的策略决策平面：将 PolicyEngine 从同步两态升级为异步三态（Allow/Deny/Ask），
引入 ApprovalBroker 处理人工审批流，并集成到 AgentLoop 的模型请求改写和工具调用执行路径。
同时修复代码审查中发现的重复代码、类型不一致、死代码、phase 映射 bug 等问题。

📁 文件变更详情
- 📄 `crates/core/src/policy/engine.rs` — 重写 PolicyEngine 为 async trait，新增 ModelRequest/CapabilityCall/PolicyVerdict/ApprovalRequest/ApprovalResolution/ContextPressureInput 等核心策略类型，Allow/Deny/Ask 三态取代原有布尔 allowed（支持审批中间态）
- 📄 `crates/core/src/policy/mod.rs` — 导出新增的策略类型与审批载荷
- 📄 `crates/core/src/lib.rs` — 更新 policy 模块公共导出列表
- 📄 `crates/core/src/action.rs` — ToolExecutionResult.duration_ms 类型从 u128 收窄为 u64（精度足够且与 StorageEvent 一致）
- 📄 `crates/core/src/capability.rs` — 为 CapabilityKind::custom() 添加文档注释说明与 new() 的等价关系
- 📄 `crates/core/src/event/translate.rs` — 修复 phase_of_storage_event 中 Error 事件的 Interrupted 分支映射，被中断会话不再错误显示为 Idle
- 📄 `crates/core/src/event/query.rs` — 消除重复的 phase 映射函数，改为调用 translate 模块的统一实现
- 🗑️ `crates/core/src/registry/capability.rs` — 删除仅含一行转发的空壳文件，直接从 registry/mod.rs 导出
- 📄 `crates/core/src/registry/mod.rs` — 移除对已删除 capability 子模块的引用
- 📄 `crates/core/src/registry/router.rs` — 新增 descriptor() 查询方法支持按名称获取 CapabilityDescriptor；duration_ms 统一为 u64
- ➕ `crates/runtime/src/approval_service.rs` — 新增 ApprovalBroker trait 与 DefaultApprovalBroker（按请求 default 值自动 resolve，避免无真实审批 transport 时 turn 卡死）
- 📄 `crates/runtime/src/lib.rs` — 注册 approval_service 模块并导出
- 📄 `crates/runtime/src/agent_loop.rs` — AgentLoop 新增 policy 和 approval 字段及 builder 方法，提供 policy_context() 构造策略上下文
- 📄 `crates/runtime/src/agent_loop/turn_runner.rs` — 在发送 LLM 请求前调用 check_model_request 进行策略改写，工具执行前调用 check_capability_call 执行三态裁决
- 📄 `crates/runtime/src/agent_loop/tool_cycle.rs` — 完整实现三态工具执行路径：Allow 直接执行、Deny 记录拒绝事件、Ask 经 ApprovalBroker 请求人工审批后再决定；新增 normalized_tool_call 校验 policy 改写不篡改 request_id 和 capability identity
- 📄 `crates/runtime/src/agent_loop/llm_cycle.rs` — generate_response 接受 ModelRequest 而非分散参数；eprintln 替换为 log::debug
- 📄 `crates/runtime/src/agent_loop/tests.rs` — 新增 4 个策略相关测试：prompt 改写、Deny 阻断执行、Ask 审批通过执行、Ask 审批拒绝阻断；新增测试辅助类型 CountingTool/RecordingApprovalBroker/RewriteSystemPromptPolicy 等
- 📄 `crates/runtime/src/service/mod.rs` — RuntimeService 持有 PolicyEngine 和 ApprovalBroker，capability reload 时保留两者而非回退默认值
- 📄 `crates/runtime/src/llm/mod.rs` — 将 HTTP client/retry/event sink 共享代码从 anthropic 和 openai 抽取到此处，新增 from_model_request 转换和 sink_collector 测试 helper
- 📄 `crates/runtime/src/llm/anthropic.rs` — 移除重复的常量、build_http_client、is_retryable_status、wait_retry_delay、emit_event、sink_collector，改为引用 llm/mod.rs 共享实现
- 📄 `crates/runtime/src/llm/openai.rs` — 同上，移除重复代码引用共享实现
- 📄 `crates/runtime/src/runtime_surface_assembler.rs` — host peer descriptor 从 "astrcode-server" 改为 "astrcode-runtime"
- 📄 `crates/runtime/src/prompt/mod.rs` — 移除多余的 #[allow(unused_imports)]
- 🗑️ `crates/plugin/src/handler_dispatcher.rs` — 删除已被 CapabilityRouter 取代的空壳 dispatcher
- 🗑️ `crates/plugin/src/transport/websocket.rs` — 删除仅有 stub 错误的 WebSocketTransport
- 📄 `crates/plugin/src/lib.rs` — 移除已删除模块的声明和导出
- 📄 `crates/plugin/src/transport/mod.rs` — 移除 websocket 模块引用
- 📄 `crates/plugin/src/invoker.rs` — duration_ms 统一为 u64
- 📄 `crates/protocol/src/plugin/descriptors.rs` — 为 CapabilityKind::custom() 添加文档注释
- 🗑️ `crates/sdk/src/agent.rs` — 删除从未使用或导出的 AgentContext
- 🗑️ `crates/sdk/src/memory.rs` — 删除无任何实现的 MemoryProvider trait
- 📄 `crates/sdk/src/lib.rs` — 移除已删除模块的声明和导出
- 📄 `crates/sdk/src/macros.rs` — declare_tool! 标记 #[doc(hidden)]
- 📄 `crates/sdk/src/stream.rs` — StreamWriter::records() 返回 Result 而非 expect panic
- 📄 `crates/sdk/src/tests.rs` — 适配 records() 返回 Result 的签名变更
- 📄 `crates/server/src/mapper.rs` — 适配 duration_ms 类型变更（u64 → u128 用于 DTO 序列化）
- 📄 `crates/tools/src/tools/edit_file.rs` — duration_ms 统一为 u64
- 📄 `crates/tools/src/tools/find_files.rs` — duration_ms 统一为 u64
- 📄 `crates/tools/src/tools/grep.rs` — duration_ms 统一为 u64
- 📄 `crates/tools/src/tools/list_dir.rs` — duration_ms 统一为 u64
- 📄 `crates/tools/src/tools/read_file.rs` — duration_ms 统一为 u64
- 📄 `crates/tools/src/tools/shell.rs` — duration_ms 统一为 u64
- 📄 `crates/tools/src/tools/write_file.rs` — duration_ms 统一为 u64
- 📄 `examples/example-plugin/src/main.rs` — 适配 records() 返回 Result 的签名变更，补齐错误映射
- 📄 `docs/adr/0005-split-policy-decision-plane-from-event-observation-plane.md` — 补充 broker 接收 turn cancellation 的设计说明
- 📄 `docs/architecture/02-core-contracts.md` — 补充当前 PolicyEngine 实现落地状态
- 📄 `docs/architecture/03-runtime-assembly.md` — 补充 approval_service.rs 模块说明和 RuntimeService 持有策略服务的说明
- 📄 `docs/architecture/04-events-approval-and-transports.md` — 补充 broker 接收 CancelToken、默认 broker 行为说明和当前实现状态
- 📄 `docs/architecture/05-refactor-roadmap.md` — 更新 Phase 3 完成状态和待后续阶段的工作项
- ➕ `docs/code-review-issues.md` — 新增代码审查问题清单及修复状态追踪

💡 变更原因
此前 PolicyEngine 仅有同步布尔 allowed/denied 两态，无法支持需要人工审批的 Ask 中间态，
也不具备对模型请求进行策略改写（如注入安全 guardrail）的能力。本次将策略系统升级为异步三态，
为后续 Web/CLI/ACP 接入真正的人工审批 transport 打下基础。同时顺带修复了代码审查发现的高优先级问题：
LLM 提供商重复代码、phase 映射 bug、duration_ms 类型不一致、多处死代码等。

⚠️ Breaking Changes
- `PolicyEngine` trait 从同步改为 async，签名不兼容，外部实现需加 `async_trait` 并改为 async fn
- `duration_ms` 类型从 `u128` 统一为 `u64`，序列化格式不变但反序列化超大值会溢出（实际不可能）

🔗 关联
Refs: Phase 3 策略决策平面落地

+++ crates/core/src/action.rs
@@ -34,7 +34,7 @@ pub struct ToolExecutionResult {
     pub error: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub metadata: Option<Value>,
-    pub duration_ms: u128,
+    pub duration_ms: u64,
     /// Indicates whether the output was truncated due to size limit
     #[serde(default)]
     pub truncated: bool,
