+++ crates/tools/src/tools/list_dir.rs
@@ -86,7 +86,7 @@ impl Tool for ListDirTool {
                 "count": entries.len(),
                 "truncated": truncated,
             })),
-            duration_ms: started_at.elapsed().as_millis(),
+            duration_ms: started_at.elapsed().as_millis() as u64,
             truncated,
         })
     }
